% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_planet_info.R
\name{get_planet_info}
\alias{get_planet_info}
\title{Get Basic Information about a Planet}
\usage{
get_planet_info(planet_name)
}
\arguments{
\item{planet_name}{The name of the planet.}
}
\value{
A data frame with basic information about the planet.
}
\description{
This function returns simple data about planets in our solar system.
}
\examples{
get_planet_info("Mars")
}
