% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_planet_position.R
\name{get_planet_position}
\alias{get_planet_position}
\title{Calculate Theoretical Position of a Planet in Its Orbit}
\usage{
get_planet_position(planet_name)
}
\arguments{
\item{planet_name}{The name of the planet.}
}
\value{
The x, y coordinates in the orbital plane.
}
\description{
This function calculates a simplified theoretical position of a planet based on its
distance from the Sun assuming a circular orbit. This is purely illustrative and assumes
the planet is at the '0 degrees' point at the time of calling.
}
\examples{
get_planet_position("Earth")
}

