% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{load_system}
\alias{load_system}
\title{Load an orbit_system from disk}
\usage{
load_system(path)
}
\arguments{
\item{path}{File path to an `.rds` file created by [save_system()].}
}
\value{
An `orbit_system` object.
}
\description{
Restores an `orbit_system` previously saved with [save_system()].
}
\examples{
\donttest{
sys <- create_system() |>
  add_sun() |>
  add_planet("Earth", parent = "Sun")

path <- file.path(tempdir(), "my_system.rds")
save_system(sys, path)
restored <- load_system(path)
}
}
