% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_orbit_simple.R
\name{plot_orbit_simple}
\alias{plot_orbit_simple}
\title{Plot a Simple Circular Orbit}
\usage{
plot_orbit_simple(radius)
}
\arguments{
\item{radius}{The radius of the orbit.}
}
\description{
This function plots a simple circular orbit to illustrate the concept of orbits.
}
\examples{
plot_orbit_simple(1)
}
