% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_orbits}
\alias{plot_orbits}
\title{Plot Orbital Trajectories (Smart 2D/3D Dispatch)}
\usage{
plot_orbits(sim_data, three_d = NULL)
}
\arguments{
\item{sim_data}{A tibble output from `simulate_system()`}

\item{three_d}{Logical. If TRUE, forces a 3D plot even for 2D data.}
}
\value{
A `ggplot` object (2D) or a `plotly` HTML widget (3D) showing the
  orbital trajectories of all bodies in the simulation.
}
\description{
Plot Orbital Trajectories (Smart 2D/3D Dispatch)
}
