% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_orbit_adjustments.R
\name{simulate_orbit_adjustments}
\alias{simulate_orbit_adjustments}
\title{Simulate Orbit Adjustments}
\usage{
simulate_orbit_adjustments(planet_data, planet_name, velocity_factor)
}
\arguments{
\item{planet_data}{A data frame containing data about planets.}

\item{planet_name}{The name of the planet to simulate.}

\item{velocity_factor}{A numeric value representing the factor by which the
planet's velocity is to be increased or decreased.}
}
\value{
A data frame containing the original and adjusted orbit distances.
}
\description{
This function simulates adjustments to a planet's orbit by changings its velocity.
The simulation is hypothetical and demonstrates how changes in velocity could
affect the orbit distance of a planet around the sun.
}
\examples{
data(planets)
simulate_orbit_adjustments(planets, "Earth", 1.1)
}
