% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_planet_position.R
\name{get_planet_position}
\alias{get_planet_position}
\title{Calculate Planet Position Using Kepler's Laws}
\usage{
get_planet_position(semi_major_axis, eccentricity, anomaly)
}
\arguments{
\item{semi_major_axis}{Semi-major axis of the planet's orbit in astronomical units.}

\item{eccentricity}{Eccentricity of the orbit.}

\item{anomaly}{Mean anomaly at a specific time in degrees.}
}
\value{
A list with x and y coordinates of the planet in its orbital plane.
}
\description{
This function calculates the position of a planet based on Kepler's laws of planetary motion.
}
\examples{
get_planet_position(1, 0.0167, 0)  # Example for Earth
}
