% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_orbits}
\alias{plot_orbits}
\title{Visualize orbital trajectories in 2D}
\usage{
plot_orbits(sim_data)
}
\arguments{
\item{sim_data}{A tidy `tibble` containing the physical state data (time, x, y, id)
generated by the `simulate()` function.}
}
\value{
A `ggplot` object representing the trajectory map.
}
\description{
Generates a 2D top-down projection of the orbital paths using `ggplot2`.
This function traces the historical trajectory of each body and places a solid
point at its final computed position. It automatically enforces a 1:1 aspect ratio
(`coord_equal`) to ensure that circular orbits are visually accurate and not distorted.
}
\examples{
\dontrun{
# Build, simulate, and plot a 28-day Earth-Moon system
orbit_data <- create_system() |>
  add_body(id = "Earth", mass = 5.97e24) |>
  add_body(id = "Moon", mass = 7.34e22, x = 3.84e8, vy = 1022) |>
  simulate(time_step = 3600, duration = 86400 * 28)

plot_orbits(orbit_data)
}
}
